feat: Increase max snapshots allowed per canister (#4169)
This increases the maximum allowed snapshots per canister from 1 to 10.
It should allow people to take multiple snapshots at different times and
be able to restore their canister to different versions if desired.

The maximum number is moved to the execution config to make it possible
to write the test in execution_environment in a more robust way that
does not depend on the hardcoded value. In the future, if we need to
change the number it should really be one line change.

---------

Co-authored-by: mraszyk <31483726+mraszyk@users.noreply.github.com>