fix: No warnings on XNet payload building
While building a XNet payload, we use the same logic to validate individual slices before including into a block as when validating slices in a proposed payload. An invalid slice in the latter case easily qualifies as a warning (a block maker has just proposed an invalid block); whereas in the former case, validation errors could simply be triggered by the replica having fallen behind subnet consensus (and thus e.g. observing signals that are ahead of the streams in its latest certified state). Meaning that during payload building it is more appropriate to reduce the log level to info.