chore(IDX): don't diff bazel targets
This drops the `diff.sh` workaround that was used to figure out which
subset of bazel targets to rebuild.

The artifact download from the cache is (now) fast enough that we want
to try avoiding the complexity of `diff.sh` and instead try to rely on
the cache for providing pre-built artifacts.