feat: CON-1422 CON-1423 Introduce VetKD payload section and deliver completed agreements  (#4022)
This PR introduces a new VetKD payload section that is filled and
validated by the new `VetKdPayloadBuilder`. The section contains
completed VetKD agreements, which are deserialized and delivered to
execution as part of the `Batch`.

The protobuf change is compatible with the mainnet version, as it only
affects payloads of data blocks (as opposed to summary blocks, which do
need to be compatible during upgrades).

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>