feat(EXC-1821): Enable on-disk compilation cache (#3554)
EXC-1821

Switch from a purely in-memory compilation cache to one that stores
wasmtime compilation artifacts on disk. This will allow us to increase
the size of the cache and reduce sandbox memory by mapping files for the
executable code instead of copying them into anonymous memory. It will
also allow us to remove the `execmem` selinux permisions for sandboxes.