test(nns): Fix tests that see less and less voting power. (#4196)
# Observed Symptom(s)

The later you run these tests, the more that voting power gets reduced.

# Cause

The clock in the test is set to the current real time, but the
timestamps in the test are hard-coded; in particular,
`voting_power_refreshed_timestamp_seconds` in Neuron was set to [the
default value]. Thus, starting on Mar 2 (in the real world), the test
starts seeing that neurons have not refreshed in a "long time" (where
"long time" is defined by [this parameter]). Thus, voting power
reduction started to kick in, leading to the observed exercised voting
power being less and less. Whereas, the expected amount of exercised
voting power is some hard-coded value. Thus, the `assert` condition
stopped being met.

[the default value]:
https://sourcegraph.com/github.com/dfinity/ic@9d4240be28224b578e58f3c1f3734a6729905226/-/blob/rs/nns/governance/src/lib.rs?L184-194

[this parameter]:
https://sourcegraph.com/github.com/dfinity/ic@9d4240be28224b578e58f3c1f3734a6729905226/-/blob/rs/nns/governance/canister/governance.did?L645-653

# Fix

Set `voting_power_refreshed_timestamp_seconds` of neurons in the test to
the current time.

# Moral of the Story

Using the current real time makes tests prone to being non-reproducible.
More generally, tests should control the situation.

One way to try to make tests that use the real time work is for all
timestamps to be relative to the current real time, which is partially
what this fix does. This is not ideal though, because it is very easy to
get this wrong. It is "much easier" to ban tests from calling `now`.
Ofc, we already have a bunch of tests that do this, and they are working
ok, so it is hard to justify going back, and making them compliant to
this suggested rule...