test: add integration tests for the socks proxy path in the outcalls adapter (#4168)
The adapter uses a socks proxy for requests iff the subnet is a system
subnet, and the direct request failed (note that a 500 code is not
consider a failure; only TCP level errors are considered failures). This
is used to connect to IPv4 targets if the node only has IPv6.

Before, this fallback behaviour was not tested. There is a [system
test](https://github.com/dfinity/ic/blob/15b9be77a0a6d3c82fdb3c187d7af3d248f3679a/rs/tests/networking/BUILD.bazel#L55)
fro it, but it's really flaky because of firewall rules / other issues
with how the BNs are setup in the farmnets.

After, there are two integration tests which cover both the old socks
proxy logic (via config), and the new ones (round robin on the addresses
in the request).

There were two main challenges:

1. **Setting up an actual socks proxy in the test**. The integration
test is quite self contained; it cannot rely on the boundary nodes to
proxy requests, thus I had to spin up a socks proxy locally. I used
`[socks5_impl](https://crates.io/crates/socks5-impl)` to help, but there
was still quite some config work to do. I am open to any suggestions if
anyone knows an easier way.
2. **Forcing the adapter to take the fallback**. There were a few
options I considered here:
- ❌ Emulating the prod behaviour: direct requests are IPv6, the target
warp server only works with IPv4. Turns out it's really hard to force
direct requests on a certain IP version, and doing any hacks might
become device dependent.
- ❌ Blocking requests at the `warp` server level. This also didn't work
because even if the target server rejects requests (from say anything
but the socks server), and it returns say 500, the request still made it
through theoretically, and hence the adapter doesn't try again.
- ✅ Forcing the socks proxy to always forward requests to a specified
URL, rather than the url in the original request, and make the direct
request to an unreachable URL. Because the URL is unreachable, this
fails at the TCP level, hence the adapter tries via the socks proxy. A
regular socks proxy would also fail, but because the test socks server
is forced to point to the "good" destination, it works.
- This required me to disable http**S** in those tests, as hyper does
not like making a TLS handshake with one server, but then connecting to
another (which makes sense).

In addition, in order to be able to test the dark launch result, I made
the adapter return it instead of the regular path result if the latter
resulted in an error. There were no cases where the main path error'd,
and the DL path succeeded in prod, so this should not influence the
behaviour too much. And if it does happen, it will be logged, and the
caller would've anyway received an error.