test(EXC-1982): Benchmark compilation cost (#4220)
EXC-1982

Add a "benchmark" in our compilation time bench suite that reports the
compilation cost of each Wasm and the expected compilation time based on
that cost. This will allow us to see how accurate the compilation costs
are.

The benchmark test cases previously had hard-coded compilation costs
which would need to be updated if the compilation costs changed. These
are no longer needed now that we have a simple way to display the
compilation costs along with the compilation time benchmark results.