refactor(CRP-2612): Renaming for ReshareChainKeys (#4131)
This is Part 1 of CRP-2612. 

It essentially contains two changes:
1. It renames `IDkgDealingsContext` to `ReshareChainKeyContext`. This
includes renaming the Protobuf type as well as the field name in the
`SubnetCallContext`. This is a backwards compatible change, as the types
don't change only their names.
2. Internally, in the type system, we were enforcing that
`IDkgDealingsContext` does not contain a VetKey, by having am
`IDkgMasterPublicKeyId` struct, that guaranteed, that we were using an
IDkgKey. Since it makes no longer sense to enforce this for
`ReshareChainKeyContext`, since it is their purpose to generalize of
IDkgKeys and VetKeys, we had to remove this contraint. To enforce that
no `ReshareChainKeyContext` with the wrong keys show up in the idkg
submodule, we introduce a new wrapper `IDkgDealingContext`, which again
guarantees that it only deals with IDKGKeys.

---------

Co-authored-by: Leo Eichhorn <99166915+eichhorl@users.noreply.github.com>