test(nns): Use a separate binary to prepare golden state to speed up experimentation (#4236)
# Why

Currently, if we have to run tests against golden NNS/SNS state many
times, each iteration will download and untar the state before doing
anything meaningful. However, if we use a separate binary to watch and
prepare for the next iteration, the download/untar time can be hidden
behind any time needed for the user to analyze and decide what to do
with the next iteration.

# What

* Refactor the download/untar operations to a library
* Use the library to implement a binary to watch a directory and untar
if the state directory is taken (by the `StateMachine`)