fix(IDX): don't run system tests with 'local' by default (#4227)
This ensures that -- by default -- system tests are not with run the
`local` strategy. When run `local`ly, system tests don't seem to get
cached across server restarts.

The `select()` which was evaluated as a boolean was truthy meaning the
`local` tag was always added. Instead, the `local` strategy is specified
as a CLI arg in the k8s CI job.