test(nns): Improve rs/nns/integration_tests (#4268)
Changes:

* `//rs/nns/integration_tests` as a library is entirely useless and
therefore removed along with all usage. It's easy to see that all the
tests are still passing even though this library target is removed from
the test canisters
* Rename `//rs/nns/integration_tests:integration_tests_test` to
`//rs/nns/integration_tests:integration_tests` since the label is now
available.
* `copy_inactive_neurons_to_stable_memory_test` is no longer needed
since the migration was done long time ago. The test is also supposed to
fail. It's just that the CI doesn't run it.
* Add `upgrade_canisters_with_golden_nns_state` to `lib.rs` so that
cargo would index it (making it easier to develop within IDEs).