fix(nns): Avoid applying approve_genesis_kyc for more than 1000 neurons (#4273)
# Why

Currently, the execution of `ApproveGenesisKyc` proposals don't have a
bound. However unlikely, it's possible for such proposals to get
through. It's not a big issue now since changing neurons have lower
instruction cost. However, after neurons are migrated to stable memory,
this can become a problem.

# What

Apply a limit of 1000 for the maximum number of neurons to set
`kyc_verified`. This should be safe as `with_neuron_mut` should take ~5M
instructions at maximum. Also, we don't expect a single kyc to apply to
1000 neurons.