perf(nns): Add benchmarks for with_neuron_mut (#4265)
# Why

Modifying a neuron (`with_neuron_mut`) can be costly with stable memory,
and we need to be conscious about this fact especially when calling it
in a loop. Adding benchmarks against it in various cases can help us
more easily make decisions about what kind of bound is reasonable.

# What

* Add benchmarks for `with_neuron_mut`
* Adjust `add_neuron` benchmarks for before/after turning on the neuron
migration