fix(sns): Ensure all topics can be set while adding new custom proposals (#4270)
This PR removes the validation rule that prevented custom proposals
being added to an SNS from having critical topics.

This contradicts the latest approach that we took for the SNS Topics
feature:

1. Disable adding new custom proposals without topics. 
2. Prompt SNS communities to set topics for their existing custom
proposals.
3. Announce the ETA for when critical topics will begin having an effect
on the criticality of the underlying topics.

Note that setting critical topics is already permitted via the
`SetTopicsForCustomProposals` proposal, but not via
`AddGenericNervousSystemFunction`.