chore: Remove the workload generator (#4143)
The workload generator is not used in any critical parts (e.g. automated
tests or release qualification) and it's not really supported. The
execution team has developed a new, improved version of a load test tool
that we are using nowadays for such work. Drop it from the repository to
keep things clean.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>