revert: "refactor(IDX): move rust-benchmarks to daily workflow" (#4293)
Reverts dfinity/ic#4159

Rust benchmark runs on dedicated machine and is excluded with monitoring
workflow runs that are in the wait queue for more then 20 minutes.
Having it on daily workflow would be OK if it wouldn't run on
`workflow_dispatch` but then maybe Crypto team does want to run them on
`workflow_dispatch`, therefore the best is if they're in it's own
workflow.