fix: retry fetching tip block (#4301)
We are seeing a high volume of error messages like:

```
Error in syncing blocks: InternalError(false, Details { error_message: Some("Tip of the chain has index 21177563 but no block fou
nd at that index!"), extra_fields: {} })
```

Which is likely caused when we hit a slightly delayed IC node after
getting the tip index from a more up-to-date one.

This PR introduces retries for fetching the tip block, so that it can
eventually succeed getting the tip block.

It also removes logging failures to get block ranges until all retries
have been attempted, cleaning the logs even further.

This was tested locally. Before the change, we were seeing the error
logs a few times per minute, with the change the error and warn logs are
gone after running for many minutes.