feat: CON-1470 Time out vetKD requests that take longer than one DKG interval (#4078)
The NiDKG transcripts used by vetKD are reshared during every DKG
interval. However, it is important that for any given request, we use
the same transcript to create and validate vetKD shares. For that
reason, we included the NiDKG ID of the transcript to be used inside of
the vetKD request context. Specifically, this is the NiDKG ID of the
"next" transcript corresponding to the height of the request.

In the worst case, a vetKD request context may be created at height
`summary_height - 1`. It is still safe to work on this request in the
subsequent interval starting with `summary_height`, as the required
NiDKG transcript is now part of the "current transcripts" in the
summary. However, should the request require even longer than the entire
interval, then the transcript will disappear entirely from the summary
at height `summary_height + dkg_interval_length + 1`. At this point,
there is no guarantee that the required transcript is still loaded by
the crypto component, which means we should reject the request.

With this PR, we therefore time out vetKD requests that take longer than
one DKG interval to complete.

Additionally, we turn the previous error codes `CanisterError` into
`SysTransient`, indicating that a retry by the user might be successful.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>