test(sns): Upgrade SNS-W before deploying SNS (and bump the canister mainnet versions) (#4334)
This PR adds a temporary step to all SNS qualification tests that
normally work with mainnet SNS-W version to first upgrade it to the
latest version on the master branch. This prepares for an urgent SNS-W
release needed to fix the SNS deployment error caused by NNS Governance
and SNS-W APIs being out of sync (namely, NNS Governance no longer sets
the neurons field, while the mainnet SNS-W still requires that).

This change is also enabling us to bump the mainnet canister versions.