feat(CRP-2612): Implement ReshareChainKey endpoint (#4177)
This is Part 2 of CRP-2612.

This PR activates the `ReshareChainKey` endpoint, which generalizes the
`ComputeIDkgDealings` endpoint to also support non IDKG chain keys such
as VetKD keys.

Since #4131, the `compute_i_idkg_dealings` endpoint internally creates a
`ReshareChainKeyContext`.
This PR extends the `ReshareChainKeyContext` with a `target_id`. The
only difference between a `ReshareChainKeyContext` generated by
`compute_i_dkg_dealings` and `reshare_chain_key` is that the latter also
initializes the `target_id` to a random number while the first one
initializes `target_id` to zero (Also old contexts are treated as if
their `target_id` is 0). Since the IDKG variant of the code does not
make use of the target_id, this is ok.

When generating a response for a `ReshareChainKeyContext`, it is
important to return the correct response type. We look at the method
name to determine, whether a `ReshareChainKeyContext` was generated
using `compute_i_dkg_dealings` or `reshare_chain_key`. This code can be
removed, once the `compute_i_dkg_dealings` endpoint will be removed
(CRP-2613).

---------

Co-authored-by: Leo Eichhorn <99166915+eichhorl@users.noreply.github.com>