chore: EXC-1752 Add TaskQueue to CanisterStateBits (#2254)
This PR adds `TaskQueue` to `CanisterStateBits`. To keep the change
backward compatible, we are adding new optional field tasks of type
`TaskQueue` to `pb::CanisterStateBits`.

In this PR we are adding support to deserialize both protobuf versions
(old and new), while we always serialize with the new one. In the
[follow-up](https://github.com/dfinity/ic/pull/4335) (planed for the
next release) we will remove(reserve) `pb::CanisterStateBits` fields
`on_low_wasm_memory_hook_status: pb::OnLowWasmMemoryHookStatus` and
`task_queue: pb::VecDequeue` and we will remove deserialization with the
old protobuf version.

---------

Co-authored-by: Dimitris Sarlis <dimitrios.sarlis@dfinity.org>