fix(icp-rosetta): [FI-1694] add watchdog thread to oversee the sync thread (#4317)
This PR introduces an implementation of a watchdog thread that monitors
heartbeats from a monitored thread and restarts it when there's no
heartbeat during a specified timeout interval.

The watchdog class is used to monitor the background synchronization
thread in ICP Rosetta, so any potential deadlock is eventually mitigated
within 10 seconds.

The watchdog implementation will later be used for ICRC1 Rosetta as
well.

Tests:
Besides unit tests for the watchdog thread itself, I've manually tested
this by introducing random delays in the sync thread (not being
committed here) and observed the successful restarts.