refactor(consensus): remote transcript delivery (#4279)
This PR does some refactoring in the batch delivery.
The reason is that the function to deliver remote dkgs back to execution
is very convoluted and adding the ability to also deliver
`ReshareChainKey` responses (CON-1416) needs significant rewriting.

So in order to not keep PRs small, this PR does the necessary refactor.
 
- Rename the function `generate_responses_to_setup_initial_dkg_calls` to
`generate_responses_to_remote_dkgs`
- Rename `TranscriptResults` to `RemoteDkgResults`
- Move logic of matching high and low threshold into methods on
`RemoteDkgResults`
- Turn it into an enum to accomodate `ReshareChainKey` responses
- Move the `generate_responses_to_remote_dkgs` into a test section in
`deliver_batches.rs` and rename to
`test_generate_setup_initial_dkg_response`
- Refactor the test
- Add an accompanying test stup for `reshare_chain_key`

---------

Co-authored-by: Leo Eichhorn <99166915+eichhorl@users.noreply.github.com>