test(IDX): only enable `flaky = True` for tests that are >= 1% flaky (#4325)
Many tests marked with `flaky = True` are not actually flaky (anymore)
or have a very low flakiness rate (< 1%). To reduce noise we remove
these `flaky = True` settings. For the ones that are actually flaky (>=
1% flakiness rate) we make sure they're marked as such and add a comment
with their actual flakiness rate to document the motivation for `flaky =
True`.

We also remove `flaky = False` settings since that's the default.