fix: Remove flakiness from the bitcoin adapter integration test (#4375)
The flaky test was `test_bfs_order`. It was setting up a 5 block long
chain, out of which 2 branches, both 6 blocks long started; meaning
there were a total of 17 blocks. Because we only send 8 blocks per
request per peer, and there are 2 peers, we needed (because 17 > 16) 2
"ticks" of requests. This eventually led to peer 1 requesting blocks
which it doesn't have, then, on the next round robin of the peers, peer
2 requesting some blocks it didn't have. Blocks from peer 1 timeout, and
it will eventually try again (because it's its turn), and the cycle only
(eventually, maybe) stops because of `Instant::elapsed()` non
determinism.

I reduced the number of total blocks in the tree to 14, so they can all
be tried in one tick, and now there won't be a divergence anymore
(whenever blocks time out, they are retried within the same tick).

In real life such a scenario has astronomically low odds of happening.
We communicate with 5 peers instead of 2, and the spread of blocks that
timeout would have to perfectly align across 5 `tick` durations (100ms).
Additionally, as 5 and 2 are co-prime, eventually the round robin will
be just right, making us send the right requests to the right peers.
Therefore, it would have to be a 5 branch situation, where somehow each
branch has at least one peer that doesn't recognize it.

A solution would be to track which peer failed to fetch which block, and
not retry that peer <-> block pair. I will implement this, but for now I
am unflaking this test, as it was affecting multiple engineers.