test(IDX): have a 30s timeout for creating the Universal and Message Canisters (#4381)
We occasionally observe that the first attempt of installing a
UniversalCanister in a system-test takes a long time and eventually
times out after 5 minutes. A subsequent attempt usually succeeds quickly
as show in the following example:
```
2025-03-14 10:21:50.735 DEBG[query_stats_above_threshold:rs/tests/driver/src/driver/test_env_api.rs:2056:0] Func="install UniversalCanister 5v3p4-iyaaa-aaaaa-qaaaa-cai [rs/tests/driver/src/util.rs:169]" is being retried for the maximum of 500s with a constant backoff of 5s
2025-03-14 10:27:00.319 DEBG[query_stats_above_threshold:rs/tests/driver/src/driver/test_env_api.rs:2078:0] Func="install UniversalCanister 5v3p4-iyaaa-aaaaa-qaaaa-cai [rs/tests/driver/src/util.rs:169]" failed on attempt 1. Error: Couldn't create canister with provisional API: The request timed out....
2025-03-14 10:27:07.556 DEBG[query_stats_above_threshold:rs/tests/driver/src/driver/test_env_api.rs:2063:0] Func="install UniversalCanister 5v3p4-iyaaa-aaaaa-qaaaa-cai [rs/tests/driver/src/util.rs:169]" succeeded after 316.820980418s on attempt 2
2025-03-14 10:27:07.556 INFO[query_stats_above_threshold:rs/tests/query_stats/lib/src/aggregation.rs:59:0] Installed Universal Canister
```

Instead of waiting for the full 5 minutes this change will install a
half minute timeout for creating a Universal or Message Canister. The
hypothesis is that with this change we'll have to wait less than 5
minutes.

Whether 30 seconds is a good timeout has to be determined
experimentally. We'll do some analysis after a week over having this
merged.