feat(sns): Define proposal criticality based on proposal topics (#4368)
This PR changes the source of truth for proposal criticality, it used to
be defined as a static function of proposal types
(`function_id_to_proposal_criticality`), and now it depends on the state
of an SNS. While native proposals have the same topics in all SNSs,
custom ones can be mapped to topics however the SNS community decides,
and this mapping can change.

This change [has been announced two weeks
ago](https://forum.dfinity.org/t/psa-sns-proposal-criticality-to-be-defined-based-on-proposal-topics/41685).