refactor: Downgrade `panic` to `debug_assert` in `Payload::take_prefix`, part 2 (#4384)
Replace `panic` with `debug_assert` plus returning an empty slice if,
after having checked that an empty slice would fit; and that the slice
we have won't fit; it turns out we somehow still have an empty slice.
The slice will eventually be replaced in the pool; or the payload
builder will call us with a higher limit; so this cannot lead to
stalling the stream.