chore(IDX): flatten upload.sh (#4396)
This moves the "skip" logic out of `upload.sh` and into `upload.bzl`.
This allows us to reuse `upload.sh` wherever necessary without the
script needed to figure out if it needs to run or not.

This'll allow us to move the upload out of the build and into a new CI
step.