test(nns): Fix wait_for_quiet test. (#4416)
The problem is that `voting_power_refreshed_timestamp_seconds` (a
recently introduced field) was not set in a test neuron. As a result,
the neuron's voting power started to be "slashed" due to Periodic
Refresh of Following/Voting Power (a new feature, directly related to
the new field). As a result, the tally did not "flip" from majority Yes
to majority No. As a result, wait for quiet did not get triggered. Thus,
the assert on original line 148 (correctly) started exploding.

# Tangent

Originally (immediately prior to this change), the exploding `assert!`
left no bread crumbs. It is practically always wrong to not attach extra
bread crumbs to an `assert!` (with `assert_eq!` this is much less of a
problem, because it at least tells you the two sides of the `==`).