feat(registry): Large response API. (#4379)
# Motivation

The motivating use case is vetKeys. However, knowledge of that feature
is NOT required to understand this change (and future changes that will
build on this).

The important thing to understand is that we want clients to be able to
fetch "large" entries. Prior to this feature, this was not possible, due
to the message size limit in the ICP (2 MiB). This feature will get
around that by spreading the desired data over multiple canister calls.

The "large" value required by vetKeys is `CatchUpPackageContents`
entries, which have keys of the form
`catch_up_package_contents_${subnet_id}`.

# Changes

This just introduces the new interface, which consists of various
`HighCapacity*` types (e.g. there was already `RegistryValue`, but now
there is also `HighCapacityRegistryValue`). Clients can begin
programming against these types, even though the Registry canister does
not yet use them. The other way around also works: when Registry starts
using the new `HighCapacity*` types, unmigrated clients will still work
as long as there are no large entries, or clients never fetch large
entries. The short explanation for how this is possible is that the new
types are wire compatible. There is a more detailed explanation in the
.proto file.

# References

Closes [NNS1-3672][tick].

[tick]: https://dfinity.atlassian.net/issues/NNS1-3672

[NNS1-3672]:
https://dfinity.atlassian.net/browse/NNS1-3672?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ