feat(sns): Move `SetTopcisForCustomProposals` under the `CriticalDappOperations` topic. (#4443)
This PR implements the last phase of the new SNS proposal criticality
feature that shifted the source of truth from a hard coded function to
the topics configuration of a given SNS.

Recently added `SetTopcisForCustomProposals` proposals enabled
categorizing custom proposals under a critical topic to become critical
for the first time. However, now that the initial batches are
categorized, its time to make the `SetTopcisForCustomProposals` proposal
critical itself, as right now it is possible to circumvent proposal
criticality by changing a proposal's topic without involving the grand
majority of votes.