feat(IDX): upload artifacts with bazel run
This moves the artifact upload out of the bazel build itself. This means
the upload can be run with `bazel run`, reducing the need for injecting
credentials into the build and allowing the upload to be re-run whenever
necessary (and e.g. not cached by Bazel), also making the upload step a
bit more declarative.