sns-testing: add sns-proposal-upvote subcommand
Users may create SNS proposals that are not automatically accepted.

To allow simulating SNS proposal adoption, 'sns-proposal-upvote' subcommand
is added to the 'sns-testing'. This command will upvote a given SNS proposal
using the same identity neurons that were used in the 'swap-complete'
subcommand to complete the SNS swap previously.

'sns-proposal-upvote' will attempt to vote on behalf of all
'Swap.minimum_participants' deterministic identities participated in the
swap omitting those that do not have neurons that are eligible for SNS voting.