chore: [NODE-1577] add a file size check for ic-os upgrade images (#4453)
This adds a test to monitor the size of compressed ICOS images, ensuring
that any significant increase is intentional. If the artifact size
exceeds a configured limit, the test will fail, prompting a review of
the change. If the increase is expected, the limit can be adjusted
accordingly. This check applies only to production variants of the ICOS
images.