chore(node): update deployment-environment argument (#4510)
The deployment-environment is used as an input to generate the node's IP
address. Current node versions sanitize this deployment-environment
input to be lowercase before deriving the IP address, but older versions
do not. This change only effects the IP address generated by older IC
image versions when run with a `testnet` deployment-environment.