chore(EXC-1837): Remove execmem permissions (#4492)
EXC-1837

Create a separate selinux domain for the compiler sandbox and remove
`execmem` permissions for the replica and canister sandbox domains.
These are no longer needed with the on disk compilation cache and they
pose a security risk because they allow processes to execute anonymous
memory.

The compiler sandbox still has `execmem` permissions because `wasmtime`
internally writes the compiled code to an executable region. But this is
much safer than having `execmem` on the canister sandbox because the
compiler sandbox doesn't execute any untrusted code (it just compiles
untrusted code).