feat: [MR-342] Remove checkpoint asynchronously (#4263)
Checkpoint removal can take 10~20 seconds on subnets with numerous
canisters and thus a large number of files in the checkpoint. A
checkpoint is deleted when the last CheckpointLayout referencing it is
dropped, which typically occurs within the `remove_states_below()` API
invoked by Consensus. If checkpoint removal takes too long, it can block
the Consensus thread, hindering progress and reducing the finalization
rate.

Checkpoint deletion consists of two steps: (1) renaming the checkpoint
to a tmp_path within the fs_tmp folder and (2) deleting the files from
tmp_path. This PR offloads the second step to a background thread,
allowing `remove_checkpoint()` to return immediately after renaming,
which usually takes only a few milliseconds. This change significantly
improves the finalization rate during checkpointing, particularly on
subnets with a large number of files.