fix(nns): Track the spawning state in the TLA model (#4525)
This makes the TLA model track the spawning state of neurons.
Previously, the model interpreted "has maturity but no stake" as equal
to spawning. This was done as to avoid storing more state and thus
potentially improve the scalability of model checking, but this is not a
correct invariant, and it led to the model finding an issue where the
cached stake would be reset when merging neurons while spawning, but
this was a false positive. The model can still be checked successfully
with this addition (i.e., it doesn't really blow up the model checking
time).

I also took the opportunity to clean up the models by factoring some
common bits out.