refactor(registry): Moved `get_type` types to Registry's `api` library crate. (#4560)
Whereas, before, these were defined in `registry.proto`.

This was really easy, because it was always the case that these were
only used in the API, never to PB-encode objects.

It should have been this way from the beginning. Not sure why I defined
them in .proto file. It might be that I thought it was the prevailing
pattern, but I now see that Registry is currently pretty inconsistent in
this regard.