test: Add some more missing compliance tests for the https outcalls feature (#4496)
The initial effort was done in https://github.com/dfinity/ic/pull/3468,
migrating most of the spec tests from haskell to rust. One can track the
effort in this
[doc](https://docs.google.com/document/d/1p55AOOvC4ORgtCpae7x7jEz8bn-LXSrP152RD9fqJn8/edit?tab=t.0).

The main missing check is checking whether cycles were refunded or not.
Spec tests in haskell had a setup `ic00viaWithCyclesRefund` of tracking
canisters that receive refunds, but there is no such thing in Rust yet
(though we can add one in the future).