chore(IDX): pass diff.sh inputs explicitly
This adapts our `diff.sh` script (used to filter out unchanged targets
on PRs) to have its inputs specified explicitly instead of relying on
environment variables.

Instead of outputting the list of targets to build, the script now only
lists the targets that have changed, and that data is used by the
`bazel-test-all` action to narrow the list of targets to build.

This means the `diff.sh` script does not need to know what targets have
to be built.