chore(IDX): inline bazel action inputs
Some action inputs like `GPG_PASSPHRASE` and `BUILDBUDDY_LINKS` should
be set for virtually every call to the bazel action. So instead of
passing it as an input every time, we read the values directly from the
bazel action.

NOTE: this means a couple more jobs will now have a CI job link.