refactor(nns): Remove topic_followee_index and USE_STABLE_MEMORY_FOLLOWING_INDEX (#4592)
# Why

The `USE_STABLE_MEMORY_FOLLOWING_INDEX` flag has been set to true on
Feb. 3, released with the proposal 135063. In this PR, the feature flag
is removed while being replaced to `true`, and because of that, the
`topic_followee_index` is no longer read, and it is also cleaned up in
this PR.

# What

* `USE_STABLE_MEMORY_FOLLOWING_INDEX` is replaced with `true`, and code
paths simplified (removing dead code, etc.)
* Benchmarks using `temporarily_disable_stable_memory_following_index`
are removed
* Calls to `temporarily_enable_stable_memory_following_index` are
removed since it's no longer necessary
* The init argument no longer has `topic_followee_index`, and the candid
files are updated accordingly
* Remove `topic_followee_index` from the canister state, along with code
to backup, restore, and maintain it (e.g. updating the index when there
are changes to the neurons
* Update benchmarks (all improvements, mostly because the code to
maintain the index is removed)

# Why it's safe

Here are the mainly reasons why it's safe to remove it from the API
level and the state:

* On the API level, the `topic_followee_index` should not have been
provided within the init argument in the first place. Any attempt to
provide a `topic_followee_index` will likely cause some inconsistency.
Also, the changes should not affect mainnet because init is not expected
to be called there.
* As part of the state, protobuf serialization/deserialization still
works when a field is removed.